Lecture 63:  Using Ternary Expressions & Comparison Operators
Instead of storing Widgets in the activeScreen variable, we are now storing strings.  This means we no longer need to rely on the lifecycle hook to make sure the switchScreen function is provided before it is passed to StartScreen.  Instead, we can now compare the string in the build method to choose which screen should be stored in the activeScreen variable.  By the time the build method is called, we have definitely received the function pointer for switchScreen that is passed to StartScreen.