Lecture 64:  Understanding "if" statements
Demonstrated a different approach to setting the active screen.  Here we're doing our comparison in the build method itself, instead of a Ternary expression.  This also avoids the lifecycle issue from a couple of lectures ago since we are doing the comparison when the widget's build method is called, which is after we receive the parameters from the calling code.