Lecture 122:  Showing and managing SnackBars
Here we used ScaffoldMessenger to create a SnackBar on the screen when the user deletes an expense entry.  The SnackBar allows the user to undo the delete action.  We also keep track of the expense's index before deletion, so we have a chance to reinsert it into the list after it has been deleted, if the user chooses to undo that action.  We're clearing any previous SnackBar in case the user deletes multiple items quickly.