Lecture 125:  Setting and using a ColorScheme
Here we are demonstrating a recommended way to theme an app.  Instead of creating an entirely new ThemeData for MaterialApp, and then configuring every property, we can simply create a ColorScheme from a seed color, and then use those colors to override things in the theme.  In this example, we've only overridden a couple things in our AppBarTheme.